+++ callisto_core/reporting/view_partials.py
@@ -325,6 +325,7 @@ class ConfirmationPartial(
             'public_key': self.coordinator_public_key,
         }
         if self.in_demo_mode:
+            kwargs['DEMO_MODE'] = True
             kwargs['to_addresses'] += self.all_user_emails
         NotificationApi.send_report_to_authority(**kwargs)
 
@@ -335,6 +336,7 @@ class ConfirmationPartial(
             'site_id': self.site_id,
         }
         if self.in_demo_mode:
+            kwargs['DEMO_MODE'] = True
             kwargs['to_addresses'] += self.all_user_emails
         NotificationApi.send_confirmation(**kwargs)